+++ cms/app/Domain/Operations/HealthCheck.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Domain\Operations;
+
+interface HealthCheck
+{
+    public function run(): HealthResult;
+}